+++ prj/vs2022/CvtOnnx.vcxproj.filters
@@ -99,6 +99,9 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertDequantizeLinearNode.cpp">
       <Filter>OnnxRuntime\D</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertDivNode.cpp">
+      <Filter>OnnxRuntime\D</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertDropoutNode.cpp">
       <Filter>OnnxRuntime\D</Filter>
     </ClCompile>
