+++ src/Cvt/OnnxRuntime/OnnxRuntime.h
@@ -61,79 +61,6 @@ namespace Synet
 
         //-----------------------------------------------------------------------------------------
 
-        bool ConvertDivNode(const onnx::NodeProto& node, const LayerParams& layers, const Bytes& original, LayerParam& layer, Bytes& reordered)
-        {
-            if (!CheckSourceNumber(layer, 2))
-                return false;
-            const LayerParam* src0 = GetLayer(layers, layer.src()[0]);
-            const LayerParam* src1 = GetLayer(layers, layer.src()[1]);
-            if (src0 == NULL || src1 == NULL)
-                return false;
-            if (src1->type() == LayerTypeConst && TensorSize(src1->weight()[0].dim()) == 1)
-            {
-                layer.type() = Synet::LayerTypePower;
-                const float* pScale = GetWeight<float>(original, src1->weight()[0]);
-                layer.power().scale() = float(1.0 / double(pScale[0]));
-                layer.src().resize(1);
-            }
-            else if (src0->type() == LayerTypeConst && TensorSize(src0->weight()[0].dim()) == 1)
-            {
-                const float* pSrc0 = GetWeight<float>(original, src0->weight()[0]);
-                if (pSrc0[0] != 1.0f)
-                    return false;
-                layer.type() = Synet::LayerTypeUnaryOperation;
-                layer.unaryOperation().type() = UnaryOperationTypeRcp;
-                layer.src().erase(layer.src().begin());
-                return true;
-            }
-            else if (src1->type() == LayerTypeConst && SignificantDimsCount(src1->weight()[0].dim()) == 1)
-            {
-                const float* pSrc = GetWeight<float>(original, src1->weight()[0]);
-                size_t size = TensorSize(src1->weight()[0].dim());
-                bool uniform = true;
-                for (size_t i = 1; i < size && uniform; ++i)
-                    uniform = (pSrc[i] == pSrc[0]);
-                if (uniform)
-                {
-                    layer.type() = Synet::LayerTypePower;
-                    layer.power().scale() = 1.0f / pSrc[0];
-                }
-                else
-                {
-                    layer.weight() = src1->weight();
-                    const Shape& dim = src1->weight()[0].dim();
-                    if ((dim.size() == 4 && dim[1] != 1) || (dim.size() == 3 && dim[0] != 1) || dim.size() == 1)
-                    {
-                        layer.type() = Synet::LayerTypeScale;
-                        layer.scale().biasTerm() = false;
-                        if (dim.size() == 1)
-                            layer.scale().axis() = -1;
-                        if (!CompactShape(layer.weight()[0].dim()))
-                            return false;                    
-                    }
-                    else
-                    {
-                        layer.type() = Synet::LayerTypeMul;
-                    }
-                    float* pDst = GetWeight<float>(reordered, layer.weight()[0]);
-                    for (size_t i = 0; i < size; ++i)
-                        pDst[i] = 1.0f / pSrc[i];
-                }
-                layer.src().resize(1);
-            }
-            else if (src0->type() == LayerTypeMeta && src1->type() == LayerTypeMeta)
-            {
-                layer.type() = Synet::LayerTypeMeta;
-                layer.meta().type() = Synet::MetaTypeDiv;
-            }
-            else
-            {
-                layer.type() = Synet::LayerTypeBinaryOperation;
-                layer.binaryOperation().type() = BinaryOperationTypeDiv;
-            }
-            return true;
-        }
-
         bool ConvertErfNode(const onnx::NodeProto& node, LayerParam& layer)
         {
             layer.type() = Synet::LayerTypeUnaryOperation;
